styling drop down is proving harder than i thought
trying to move it with margin seems to leave it unresponsive to my
movement commands. Also adding to the total height only further worsens
its current positioning. I did however manage to style the SELECT a bit
and change the font and cool things and removed border as well , trying
to use the CSS appearance property too